Clarify that tokens are minted into user's token accounts
Heya! I'm contributing here because my original impression from reading this doc was a little misleading. When I read 

"Every token mint has a different token account associated with it."

I thought this meant there was a 1:1 relationship of mints to token accounts. Eg, FooToken mint account would mint to FooToken token account, and then be transferred to some user's FooToken token account.

From what I've been told, that's incorrect - [talking with others in Solana Stack Exchange](https://solana.stackexchange.com/questions/849/are-new-spl-tokens-typically-transferred-from-a-mints-token-account-to-a-users), it seems that a token is minted directly to a user's token account. FooToken mint account, mints token to Joe Smith's FooToken Token Account.

Is that correct? If so I've made some changes to clarify this, which should help clarify things for other people learning Solana.